fix: remove erroneous GET argument from curl call
This change removes a stray `GET` argument that was being passed to the curl command in the `list-all` script. At least with curl v8.5.0, it was taking ages from this script to run because curl first trying to hit the address `GET` (equivalent `curl --silent GET`) and waiting on it for quite some time before actually moving on to hitting the gcs bucket url.

Ultimately, this was causing my terminal to hang for ages when I had `asdf:mise-plugins/mise-gcloud` in my mise tools.